+++ src/Discord.Net.Models/Interfaces/IOverwriteModel.cs
@@ -1,9 +1,11 @@
+using System.Numerics;
+
 namespace Discord.Models;
 
 public interface IOverwriteModel : IEntityModel
 {
     ulong TargetId { get; }
     int Type { get; }
-    ulong Allow { get; }
-    ulong Deny { get; }
+    BigInteger Allow { get; }
+    BigInteger Deny { get; }
 }
