+++ src/Discord.Net.Models/RestModels/Guilds/Overwrite.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using System.Text.Json.Serialization;
 
 namespace Discord.Models.Json;
@@ -11,8 +12,8 @@ public sealed class Overwrite : IOverwriteModel
     public int Type { get; set; }
 
     [JsonPropertyName("deny")]
-    public ulong Deny { get; set; }
+    public BigInteger Deny { get; set; }
 
     [JsonPropertyName("allow")]
-    public ulong Allow { get; set; }
+    public BigInteger Allow { get; set; }
 }
